+++ crates/winnow-datetime/src/macros.rs
@@ -6,7 +6,7 @@ macro_rules! duration_part_seq {
         fraction: $fraction_parser:expr,
         end: $end_parser:expr
     }) => {
-        move |input: &mut _| {
+        (move |input: &mut _| {
             seq!((
                 $whole_parser,
                 opt(preceded($sep_parser, $fraction_parser.map(|n| {
@@ -17,7 +17,7 @@ macro_rules! duration_part_seq {
             ))
             .map(move |(whole, frac)| DurationPart{ whole, frac })
             .parse_next(input)
-        }
+        })
     };
 }
 
@@ -28,13 +28,13 @@ macro_rules! date_yddd_seq {
         year: $year_parser:expr,
         day: $day_parser:expr,
     }) => {
-        move |input: &mut _| {
+        (move |input: &mut _| {
             seq!($date_type::$variant {
                 year: $year_parser,
                 day: $day_parser,
             })
             .parse_next(input)
-        }
+        })
     };
 }
 
@@ -77,7 +77,7 @@ macro_rules! time_seq {
         hour: $hour_parser:expr,
         minute: $minute_parser:expr,
         second: $second_parser:expr,
-        millisecond: $millisecond_parser:expr,
+        nanosecond: $nanosecond_parser:expr,
         offset: $offset_parser:expr,
         time_zone: $time_zone_parser:expr,
         calendar: $calendar_parser:expr,
@@ -86,7 +86,7 @@ macro_rules! time_seq {
             hour: $hour_parser,
             minute: $minute_parser,
             second: $second_parser,
-            millisecond: $millisecond_parser,
+            nanosecond: $nanosecond_parser,
             offset: $offset_parser,
             time_zone: $time_zone_parser,
             calendar: $calendar_parser,
@@ -101,14 +101,14 @@ macro_rules! partial_time_seq {
         hour: $hour_parser:expr,
         minute: $minute_parser:expr,
         second: $second_parser:expr,
-        millisecond: $millisecond_parser:expr,
+        nanosecond: $nanosecond_parser:expr,
         offset: $offset_parser:expr,
     }) => {
         seq!($variant {
             hour: $hour_parser,
             minute: $minute_parser,
             second: $second_parser,
-            millisecond: $millisecond_parser,
+            nanosecond: $nanosecond_parser,
             offset: $offset_parser,
         })
     };