+++ crates/winnow-iso8601/src/fractional_duration.rs
@@ -9,9 +9,12 @@ use winnow::token::literal;
 use winnow::{seq, Parser, Result};
 use winnow_datetime::FractionalDuration;
 
-/// Parses a duration with the same formating rules but allows for decimal places.
+/// Parses a duration with the same formatting rules but allows for decimal places.
+///
+/// ```rust
 /// let duration = winnow_iso8601::parse_fractional_duration("P1,5Y2M3DT4,5H5M6S").unwrap();
 /// let duration = winnow_iso8601::parse_fractional_duration("P1,5W").unwrap();
+/// ```
 pub fn parse_fractional_duration(mut i: &str) -> Result<FractionalDuration, InputError<&str>> {
     terminated(fractional_duration, eof).parse_next(&mut i)
 }
@@ -19,7 +22,7 @@ pub fn parse_fractional_duration(mut i: &str) -> Result<FractionalDuration, Inpu
 /// Parses a duration string with the format P%dY%dM%dDT%dH%dM%dS
 pub fn fractional_duration<'i, Input, Error>(
     input: &mut Input,
-) -> std::result::Result<FractionalDuration, Error>
+) -> core::result::Result<FractionalDuration, Error>
 where
     Input: StreamIsPartial + Stream + Compare<&'i str>,
     <Input as Stream>::Slice: AsBStr,
@@ -36,11 +39,7 @@ where
             opt(preceded(opt(literal("T")), duration_base_time)),
         ))
         .verify(|(y, mo, w, d, time)| {
-            if y.is_none() && mo.is_none() && w.is_none() && d.is_none() && time.is_none() {
-                false
-            } else {
-                true
-            }
+            !(y.is_none() && mo.is_none() && w.is_none() && d.is_none() && time.is_none())
         })
         .map(|(y, mo, w, d, time)| {
             let time = time.unwrap_or((None, None, None));