+++ crates/winnow-iso8601/src/time.rs
@@ -4,7 +4,7 @@ use winnow::error::{InputError, ParserError};
 use winnow::stream::{AsBStr, AsChar, Compare, Stream, StreamIsPartial};
 use winnow::token::{literal, one_of};
 use winnow::{seq, Parser, Result};
-use winnow_datetime::parser::{fraction_millisecond, time_hour, time_minute, time_second};
+use winnow_datetime::parser::{fraction_nanosecond, time_hour, time_minute, time_second};
 use winnow_datetime::Time;
 
 /// Parses a time string.
@@ -29,7 +29,7 @@ pub fn parse_time(mut i: &str) -> Result<Time, InputError<&str>> {
 /// * `0735[00][.123][(Z|(+|-)0000)]`
 ///
 // HH:MM:[SS][.(m*)][(Z|+...|-...)]
-pub fn time<'i, Input, Error>(input: &mut Input) -> std::result::Result<Time, Error>
+pub fn time<'i, Input, Error>(input: &mut Input) -> core::result::Result<Time, Error>
 where
     Input: StreamIsPartial + Stream + Compare<&'i str>,
     <Input as Stream>::Slice: AsBStr,
@@ -51,7 +51,7 @@ where
 ///
 /// See [`time()`][`crate::time()`] for the supported formats.
 // HH:MM:[SS][.(m*)][(Z|+...|-...)]
-pub(crate) fn base_time<'i, Input, Error>(input: &mut Input) -> std::result::Result<Time, Error>
+pub(crate) fn base_time<'i, Input, Error>(input: &mut Input) -> core::result::Result<Time, Error>
 where
     Input: StreamIsPartial + Stream + Compare<&'i str>,
     <Input as Stream>::Slice: AsBStr,
@@ -65,14 +65,14 @@ where
 
         let offset = opt(offset).parse_next(input)?;
 
-        let (minute, second, millisecond) = msms.unwrap_or((0, None, None));
+        let (minute, second, nanosecond) = msms.unwrap_or((0, None, None));
 
         Ok(Time {
-            hour,                                  // HH
-            minute,                                // MM
-            second: second.unwrap_or(0),           // [SS]
-            millisecond: millisecond.unwrap_or(0), // [.(m*)]
-            offset,                                // [(Z|+...|-...)]
+            hour,                                // HH
+            minute,                              // MM
+            second: second.unwrap_or(0),         // [SS]
+            nanosecond: nanosecond.unwrap_or(0), // [.(m*)]
+            offset,                              // [(Z|+...|-...)]
             time_zone: None,
             calendar: None,
         })
@@ -83,7 +83,7 @@ where
 /// Parses secondary portion of a time string.
 pub(crate) fn time_minute_second_millisecond<'i, Input, Error>(
     input: &mut Input,
-) -> std::result::Result<(u32, Option<u32>, Option<u32>), Error>
+) -> core::result::Result<(u32, Option<u32>, Option<u32>), Error>
 where
     Input: StreamIsPartial + Stream + Compare<&'i str>,
     <Input as Stream>::Slice: AsBStr,
@@ -96,7 +96,7 @@ where
             seq!(
                 preceded(opt(literal(":")), time_minute),
                 opt(preceded(opt(literal(":")), time_second)),
-                opt(preceded(one_of(b",."), fraction_millisecond))
+                opt(preceded(one_of(b",."), fraction_nanosecond))
             )
             .parse_next(input)
         },