+++ crates/winnow-rfc9557/src/time.rs
@@ -10,7 +10,7 @@ use winnow::stream::{AsBStr, AsChar, Compare, Stream, StreamIsPartial};
 use winnow::token::literal;
 use winnow::token::one_of;
 use winnow::{seq, Parser, Result};
-use winnow_datetime::parser::fraction_millisecond;
+use winnow_datetime::parser::fraction_nanosecond;
 use winnow_datetime::parser::time_hour;
 use winnow_datetime::parser::time_minute;
 use winnow_datetime::parser::time_second;
@@ -51,16 +51,11 @@ where
             hour: time_hour,                             // HH
             minute: preceded(literal(":"), time_minute), // MM
             second: preceded(literal(":"), time_second), // [SS]
-            millisecond: opt(preceded(one_of(b",."), fraction_millisecond)).map(|d| d.unwrap_or(0)), // [.(m*)]
-            offset: offset.map(|o| Some(o)),  // [(Z|+...|-...)]
+            nanosecond: opt(preceded(one_of(b",."), fraction_nanosecond)).map(|d| d.unwrap_or(0)), // [.(m*)]
+            offset: offset.map(Some),         // [(Z|+...|-...)]
             time_zone: opt(suffix_time_zone), // [time zone]
-            calendar: opt(repeat(1.., suffix_calendar)).map(|c: Option<Vec<Calendar>>| {
-                if let Some(c) = c {
-                    Some(c[0].clone())
-                } else {
-                    None
-                }
-            }),
+            calendar: opt(repeat(1.., suffix_calendar))
+                .map(|c: Option<Vec<Calendar>>| { c.map(|c| c[0].clone()) }),
         })
         .parse_next(input)
     })
@@ -153,7 +148,7 @@ mod parsers {
             hour: 2,
             minute: 4,
             second: 28,
-            millisecond: 0,
+            nanosecond: 0,
             offset: Some(Offset::LocalUnknown { critical: false }),
             time_zone: Some(TimeZone::Named {
                 zone: NamedTimeZone {